Interval Summary Statistics

Interval Summary Statistics

interval_stats(df, type = "diurnal", seas = NULL)

Arguments

df
a dataframe containing columns "x", "y", and "date"

type
a character string indicating the intervals to calculate, one of "diurnal", "lunar", or "seasonal".

seas
a character vector including the start date of each season interval in the format %Y-%m-%d, e.g. "2015-01-01". Required if type == "seasonal".

Details

Note: if your trajectory includes dates before the first date in your seas vector, they will automatically be considered a separate season.

See also

Examples

# diurnal is default
interval_stats(AG195)

#> # A tibble: 730 x 9
#> # Groups:   interval_start [730]
#>    interval_start      TOD   mean_dist sd_dist acf_dist mean_ang sd_ang acf_ang
#>    <dttm>              <chr>     <dbl>   <dbl>    <dbl>    <dbl>  <dbl>   <dbl>
#>  1 2010-01-01 00:00:00 0:00…      88.0   114.    0.515   -0.0457  1.37  -0.399 
#>  2 2010-01-01 12:00:00 12:0…     146.    181.    0.570   -0.290   1.19   0.0298
#>  3 2010-01-02 00:00:00 0:00…     147.    177.    0.591    0.493   1.31  -0.201 
#>  4 2010-01-02 12:00:00 12:0…     157.    154.    0.612   -0.0942  0.923 -0.412 
#>  5 2010-01-03 00:00:00 0:00…      87.2   125.    0.436    0.0130  0.902  0.0815
#>  6 2010-01-03 12:00:00 12:0…      43.5    64.3   0.452    0.257   1.33   0.0269
#>  7 2010-01-04 00:00:00 0:00…      70.6    96.2   0.595   -0.314   1.54   0.123 
#>  8 2010-01-04 12:00:00 12:0…      42.5    33.8   0.143   -0.0764  1.45  -0.0202
#>  9 2010-01-05 00:00:00 0:00…      95.2   111.    0.473   -0.0782  1.37  -0.102 
#> 10 2010-01-05 12:00:00 12:0…      63.3    42.6   0.0695   0.0619  1.02   0.0658
#> # … with 720 more rows, and 1 more variable: ccf <dbl>

interval_stats(AG195, type = "lunar")

#> # A tibble: 25 x 9
#> # Groups:   interval_start [25]
#>    interval_start phase mean_dist sd_dist acf_dist mean_ang sd_ang  acf_ang
#>    <fct>          <chr>     <dbl>   <dbl>    <dbl>    <dbl>  <dbl>    <dbl>
#>  1 2010-01-01     Full…      99.8    126.    0.571 -5.92e-4   1.31 -0.0351 
#>  2 2010-01-12     New-…     118.     122.    0.517  1.52e-2   1.22 -0.0611 
#>  3 2010-01-26     Full…     181.     239.    0.686 -3.61e-2   1.32  0.0199 
#>  4 2010-02-10     New-…     244.     284.    0.710  6.39e-2   1.30  0.0235 
#>  5 2010-02-25     Full…     210.     254.    0.674  6.49e-2   1.37  0.0434 
#>  6 2010-03-12     New-…     192.     261.    0.700  1.24e-2   1.37 -0.0126 
#>  7 2010-03-27     Full…     105.     152.    0.709  5.35e-2   1.28 -0.00154
#>  8 2010-04-10     New-…     184.     246.    0.751 -2.38e-2   1.21 -0.00632
#>  9 2010-04-25     Full…     119.     161.    0.539  4.81e-2   1.23 -0.119  
#> 10 2010-05-10     New-…     115.     169.    0.612  4.37e-2   1.19 -0.0617 
#> # … with 15 more rows, and 1 more variable: ccf <dbl>

# for seasonal, include y-m-d formatted `seas` vector
interval_stats(AG195, type = "seasonal", seas = c("2010-03-01", "2010-06-01", "2010-9-01"))

#> # A tibble: 4 x 8
#>   interval_start mean_dist sd_dist acf_dist  mean_ang sd_ang  acf_ang      ccf
#>   <fct>              <dbl>   <dbl>    <dbl>     <dbl>  <dbl>    <dbl>    <dbl>
#> 1 2009-12-31         174.     226.    0.697  0.0102     1.28 -0.00832  0.00843
#> 2 2010-03-01         144.     207.    0.690  0.0233     1.28 -0.0282  -0.00724
#> 3 2010-06-01          93.6    137.    0.564 -0.000686   1.25 -0.0124   0.0107 
#> 4 2010-09-01         109.     172.    0.675  0.0118     1.33 -0.0358  -0.00321